modified:   rgb_led_toy_test.h modified:   rgb_led_toy_test.pde
wobble() was replaced with wobble2(). wobble() is kept in the code in a block for obsolete functions for reference only.

wobble2() now takes a pointer to an array of uint8_t and the array's length as an uint8_t as well.

the array should contain the data for CW mode. the pattern for CCW mode is calculated automatically.