modified:   rgb_led_toy_test.h modified:   rgb_led_toy_test.pde
Adapted the code for the new LEDs from JELEDHK. The chips are reversed,
therefore the LEDs need to be rotated by 180°, which switches RED_A and GREEN_A.
For these new LEDs the silkscreen is wrong. The LEDs' polarity marks must show
towards the ATmega chip and not to the outside.